refectory: 코드 분리
